🐛 修改float32变成64

+++ models/config/tag.go
@@ -5,7 +5,7 @@ import "time"
 type Tag struct {
     Quality    int8      `json:"quality" xorm:"-"`
     Id         int64     `json:"id" xorm:"autoincr pk INTEGER"`
-    Value      float32   `json:"value" xorm:"-"`
+    Value      float64   `json:"value" xorm:"-"`
     Time       int64     `json:"time" xorm:"-"`
     CreateTime time.Time `json:"createTime" xorm:"created"`
     Code       string    `json:"code"`
